+++ client/src/main/java/ui/data/GameListAccessor.java
@@ -1,4 +1,4 @@
-package ui;
+package ui.data;
 
 import model.GameData;
