+++ client/src/test/java/client/ServerFacadeTests.java
@@ -6,8 +6,8 @@ import model.AuthData;
 import model.GameData;
 import org.junit.jupiter.api.*;
 import server.Server;
-import ui.HttpFailureException;
-import ui.ServerFacade;
+import ui.data.HttpFailureException;
+import ui.data.ServerFacade;
 import util.SerializerFactory;
 import java.io.IOException;
 import java.util.ArrayList;
